server/requirements.txt
python-telegram-bot==20.7
google-generativeai==0.8.3
pdfplumber==0.11.0
reportlab==4.1.0
python-docx==1.1.0
aiofiles==23.2.1
python-dotenv==1.0.0
Pillow==10.3.0
flask==3.0.3
gunicorn==22.0.0
flask-cors==4.0.1
